UBIFS: use time granularity consistently

+++ fs/ubifs/super.c
@@ -110,11 +110,11 @@ struct inode *ubifs_iget(struct super_block *sb, unsigned long inum)
 	inode->i_nlink = le32_to_cpu(ino->nlink);
 	inode->i_uid   = le32_to_cpu(ino->uid);
 	inode->i_gid   = le32_to_cpu(ino->gid);
-	inode->i_atime.tv_sec  = le32_to_cpu(ino->atime.sec);
+	inode->i_atime.tv_sec  = (int32_t)le32_to_cpu(ino->atime.sec);
 	inode->i_atime.tv_nsec = le32_to_cpu(ino->atime.nsec);
-	inode->i_mtime.tv_sec  = le32_to_cpu(ino->mtime.sec);
+	inode->i_mtime.tv_sec  = (int32_t)le32_to_cpu(ino->mtime.sec);
 	inode->i_mtime.tv_nsec = le32_to_cpu(ino->mtime.nsec);
-	inode->i_ctime.tv_sec  = le32_to_cpu(ino->ctime.sec);
+	inode->i_ctime.tv_sec  = (int32_t)le32_to_cpu(ino->ctime.sec);
 	inode->i_ctime.tv_nsec = le32_to_cpu(ino->ctime.nsec);
 	inode->i_mode  = le32_to_cpu(ino->mode);
 	inode->i_size  = le64_to_cpu(ino->size);
